code_1/requirements.txt
# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Text Processing & Chunking
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# Embeddings
openai>=1.0.0
sentence-transformers>=2.2.0

# Vector Store
chromadb>=0.4.0
faiss-cpu>=1.7.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
tqdm>=4.65.0
